Update pagebreaks.js
Allow for creation of any type of HTML element to fill space between pages. Allow for styling of this element, this is especially important where for example one would like to remove borders on the added area (since its in a table/tr) but cannot do that with a div.